api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API/backend for portfolio website"
authors = ["Joe Bonneau <joebonneau@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.0"
black = "^21.8b0"
mypy = "^0.910"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
